Implement router for reviews route
