+++ index.js
@@ -4,11 +4,8 @@ const methodOverride = require('method-override');
 const mongoose = require('mongoose');
 const path = require('path');
 const PORT = 3000;
-const { reviewSchema } = require('./schemas');
-const Campground = require('./models/campground');
-const Review = require('./models/review');
 const campgroundRoutes = require('./routes/campgrounds');
-const catchAsync = require('./utils/catchAsync');
+const reviewRoutes = require('./routes/reviews');
 const ExpressError = require('./utils/ExpressError');
 
 mongoose
@@ -25,44 +22,12 @@ app.set('views', path.join(__dirname, 'views'));
 app.use(express.urlencoded({ extended: true }));
 app.use(methodOverride('_method'));
 
-const validateReview = (req, res, next) => {
-  const { error } = reviewSchema.validate(req.body);
-  if (error) {
-    const message = error.details.map(detail => detail.message).join(',');
-    throw new ExpressError(message, 400);
-  } else {
-    next();
-  }
-};
-
 app.use('/campgrounds', campgroundRoutes);
+app.use('/campgrounds/:id/reviews', reviewRoutes);
 
 app.get('/', (req, res) => {
   res.render('home');
 });
-
-app.post(
-  '/campgrounds/:id/reviews',
-  validateReview,
-  catchAsync(async (req, res) => {
-    const campground = await Campground.findById(req.params.id);
-    const review = new Review(req.body.review);
-    campground.reviews.push(review);
-    await review.save();
-    await campground.save();
-    res.redirect(`/campgrounds/${campground._id}`);
-  })
-);
-
-app.delete(
-  '/campgrounds/:id/reviews/:reviewId',
-  catchAsync(async (req, res) => {
-    const { id, reviewId } = req.params;
-    await Campground.findByIdAndUpdate(id, { $pull: { reviews: reviewId } });
-    await Review.findByIdAndDelete(reviewId);
-    res.redirect(`/campgrounds/${id}`);
-  })
-);
 
 app.all('*', (req, res, next) => {
   next(new ExpressError('Page Not Found', 404));
